requirements.txt
pyrogram
tgcrypto
requests
hachoir
pytz
hachoir
ffmpeg-python